backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
whitenoise==6.7.0
